Refactoring: Add PasskeySignerProtocol
